Introduce Resource model, refactor default middleware stack
Matching Resources are categorized and given to middleware to
determine best match and implement resolve algorithm.

Biggest TODOs from here are to refactor out file serving code in the
different serve-* middleware, perhaps by adding a `MATCH_EXACT`
Resource in the case of a found index or renderable. Will need a new
match type for generated content or serve directly from that middleware.